metrics/influxdb: align v1 and v2
